Smooth topography before interp to MPAS
This merge adds a python script for iteratively smoothing
topograpthy (from for now, the dataset combining BedMachine and GEBCO)
proportional to the resolution of the MPAS mesh, as determined in
the base_mesh step.

Local weights for the iterative smoothing are determined based on
the resolution of the lon/lat grid, the MPAS mesh resolution, the
number of iterations of smoothing and a constant multiplicative
factor `smoothing` that can be increased or decreased to provide
more or less smoothing overall.  The default value,
`smoothing = 1.0`, smooths the topography by one MPAS cell width.

Increasing the number of iterations will typically not increase
the amount of smoothing, since less smoothing is performed at each
iteration.  The exception is that the maximum amount of smoothing
is an unweighted nearest-neighbor averaging at each iteration.  If
a region (commonly the poles) is so stretched on the lon/lat grid
that an MPAS cell covers more grid cells than the number of
smoothing iterations, it is not possible to smooth over the full
width of an MPAS grid cell.  In practice, this is not likely to
be an important issue.

The cost of this approach is significant, taking several minutes
even for coarse MPAS meshes (it is performed on the topographic
grid, so is largely agnostic to the MPAS resolution).  We may
need to make the QU240wISC test case a special case where we
either skip this process or simplify it in some way to reduce
runtime of regression testing.

Currently, the approach is only being used for test cases with
ice-shelf cavities (the same ones using BedMachine/GEBCO
bathymetry).  It could be expanded to other test cases if it
proves useful.