Add support for cached output files
When constructing a step, if ``cached=True``, the outputs for
this step will be downloaded to the appropriate local database and
symlinked instead of being computed.  Inputs (other than the cached
outputs) and the run method will be skipped.

Each MPAS core can optionally have a database (a python dictionary
in a json file called cached_files.json) that keeps track of which
files are available in the cache and what date stamp is in the
filename.

When setting up test cases, a user can supply test-case numbers
with a "c" suffix to indicate that they should be cached.  When
setting up test cases individually with a path, a user can supply
a list of steps in the test case that should use cached outputs.
A test suite can supply a line with "cached" or
"cached: <step> <step>" to indicate either that all steps in the
test case or the listed steps should use cached outputs.