Reorganize ice_shelf_2d into more test cases
There is a separate init test case, an ssh-adjustment test case
from the ocean framework, and some reogranization of the existing
default and restart_test test cases.

To share the ssh_adjustment test case between the two forward
test cases, we need to use the same number of iterations of ssh
adjustment between the 2.  It is more convenient to use the 2
iterations prescribed in the restart test because that test is
used in regression suites.